+++ examples/initialize_presale/initialize_fixed_price_presale.ts
@@ -42,6 +42,7 @@ async function initializeFixedPricePresale(
     {
       price,
       rounding: Rounding.Down,
+      disableWithdraw: false,
     }
   );
 
@@ -104,12 +105,14 @@ const presaleArgs: Omit<IPresaleArgs, "presaleMode"> = {
   presaleEndTime: new BN(Math.floor(Date.now() / 1000 + 3600)),
   whitelistMode: WhitelistMode.Permissionless,
   unsoldTokenAction: UnsoldTokenAction.Refund,
+  disableEarlierPresaleEndOnceCapReached: false,
 };
 
 const lockedVestingArgs: ILockedVestingArgs = {
   lockDuration: new BN(0),
   vestDuration: new BN(0),
   immediateReleaseBps: new BN(10_000),
+  immediateReleaseTimestamp: presaleArgs.presaleEndTime,
 };
 
 const baseMintPubkey = new PublicKey(